@@ -25,7 +25,7 @@
   },
   "license": "AGPL-3.0",
   "peerDependencies": {
-    "puppeteer": "^1.5.0 || ^2.0.0 || ^3.0.0 || ^4.0.0 || ^5.0.0"
+    "puppeteer": "^3.0.0 || ^4.0.0 || ^5.0.0"
   },
   "devDependencies": {
     "@types/css-tree": "^1.0.3",
@@ -34,30 +34,29 @@
     "@types/memoizee": "^0.4.4",
     "@types/minimist": "^1.2.0",
     "@types/node": "^14.0.13",
-    "@types/node-fetch": "^2.5.7",
-    "@types/puppeteer": "^3.0.1",
-    "@types/uuid": "^7.0.2",
+    "@types/puppeteer": "^5.4.2",
+    "@types/uuid": "^8.3.0",
     "characterset": "^1.3.0",
     "coveralls": "^3.1.0",
     "css-tree": "^1.0.0-alpha.39",
-    "fastify": "^2.15.0",
-    "fastify-static": "^2.7.0",
-    "jest": "^26.1.0",
-    "puppeteer": "^5.2.1",
-    "ts-jest": "^26.1.3",
-    "ts-node": "^8.8.2",
-    "ts-node-dev": "^1.0.0-pre.44",
-    "typescript": "^3.8.3",
-    "xo": "^0.28.2",
     "debug": "^4.1.1",
-    "minimist": "^1.2.5"
+    "fastify": "^3.9.2",
+    "fastify-static": "^3.3.0",
+    "jest": "^26.6.3",
+    "minimist": "^1.2.5",
+    "puppeteer": "^5.5.0",
+    "ts-jest": "^26.1.3",
+    "ts-node": "^9.1.1",
+    "ts-node-dev": "^1.0.0-pre.44",
+    "typescript": "^4.2.0-dev.20201230",
+    "xo": "^0.36.1"
   },
   "dependencies": {
     "abort-controller": "^3.0.0",
     "characterset": "^1.3.0",
     "css-tree": "^1.0.0-alpha.39",
     "memoizee": "^0.4.14",
-    "node-fetch": "^2.6.0",
+    "node-fetch": "^3.0.0-beta.9",
     "parse-cache-control": "^1.0.1"
   },
   "xo": {
