@@ -1,16 +1,14 @@
 import Collect from './collect';
-import {PrivateSettings} from '../types/settings';
+import { PrivateSettings } from '../types/settings';
 import * as util from '../utils/utils';
-import {CollectMeta} from '../types/audit';
-import {PageContext} from '../types';
-import {CollectAnimationsTraces, SingleAnimationFormat} from '../types/traces';
-// Const TraceToTimelineModel = require('devtools-timeline-model');
+import { CollectMeta } from '../types/audit';
+import { PageContext } from '../types';
+import { CollectAnimationsTraces, SingleAnimationFormat } from '../types/traces';
 
 /**
  * @overview: Get CSSTransitions/CSSAnimations, stuff that requires CPU process and see if they are stoped when tab is switched
  */
 
-// fails on : https://codebeautify.org/jsonminifier
 export default class CollectAnimations extends Collect {
 	static get meta() {
 		return {
@@ -25,10 +23,9 @@ export default class CollectAnimations extends Collect {
 		settings: PrivateSettings
 	): Promise<CollectAnimationsTraces | undefined> {
 		try {
-			// https://chromedevtools.github.io/devtools-protocol/tot/DOM/#type-Node
 			const debug = CollectAnimations.meta.debug;
 			debug('running');
-			const {page} = pageContext;
+			const { page } = pageContext;
 
 			const client = await page.target().createCDPSession();
 			await client.send('Animation.enable');
@@ -59,7 +56,7 @@ export default class CollectAnimations extends Collect {
 				const classname = getClassOrIdName(nodeAttributes, 'class');
 				const idname = getClassOrIdName(nodeAttributes, 'id');
 				if (classname || idname) {
-					data.animation.meta = {classname, idname};
+					data.animation.meta = { classname, idname };
 					animations.set(data.animation.id, data.animation);
 				}
 			});
@@ -81,7 +78,7 @@ export default class CollectAnimations extends Collect {
 							const data = animation[1];
 							const id = animation[0];
 							if (reactiveAnimationArray.includes(id)) return true;
-							const {classname, idname} = data.meta;
+							const { classname, idname } = data.meta;
 							if (!notReactiveAnimationsSet.has(data.name)) {
 								notReactiveAnimations.push({
 									name: data.name,
@@ -105,86 +102,12 @@ export default class CollectAnimations extends Collect {
 			debug('done');
 			return {
 				animations: hasAnimations
-					? {notReactive: notReactiveAnimations}
+					? { notReactive: notReactiveAnimations }
 					: undefined
 			};
-
-			/*
-
-			*/
-
-			// get initial trace
-			/* const categories: string[] = [
-				'-*',
-				'v8.execute',
-				'blink.user_timing',
-				'latencyInfo',
-				'devtools.timeline',
-				'disabled-by-default-devtools.timeline',
-				'disabled-by-default-devtools.timeline.frame',
-				'toplevel',
-				'blink.console',
-				'disabled-by-default-devtools.timeline.stack',
-				'disabled-by-default-devtools.screenshot',
-				'disabled-by-default-v8.cpu_profile',
-				'disabled-by-default-v8.cpu_profiler',
-				'disabled-by-default-v8.cpu_profiler.hires'
-			];
-
-			const processTracingAndGetSummary = async () => {
-				await page.tracing.start({
-					path: '',
-					categories,
-					screenshots: false
-				});
-				await new Promise(resolve => setTimeout(() => resolve(), 1000)); // Wait 1sec
-				const rawData = await page.tracing.stop();
-				const data = JSON.parse(rawData.toString());
-				const model = new TraceToTimelineModel(data.traceEvents);
-				const timelineModel = model.timelineModel();
-				const start = timelineModel._minimumRecordTime;
-				const end = timelineModel._maximumRecordTime;
-				// @ts-ignore
-				return util.getSummary(timelineModel, start, end);
-			};
-
-			/**
-			 * Find ROIs
-			 * initial summary
-			 * move viewport (scroll by y) outside ROI
-			 * second summary
-			 * compare diff
-			 */
-			/*
-			const getSummary = async () => {
-				const summaryArray: AnimationsFormat[] = [];
-				debug('getting init page tracing');
-				const initSummary = await processTracingAndGetSummary();
-				await new Promise(resolve => {
-					// @ts-ignore custom event
-					page.on('scrollFinished', async () => {
-						debug('getting after-scroll page tracing');
-						const endSummary = await processTracingAndGetSummary();
-
-						summaryArray.push({initSummary, endSummary});
-						resolve();
-					});
-				});
-
-				return summaryArray;
-			};
-
-			const summary = await getSummary();
-
-			page.removeListener('scrollFinished', () => {
-				debug('removed scrollFinished event listener');
-			});
-
-			debug('done');
-			*/
 		} catch (error) {
 			util.log(`Error: Animations collect failed with message: ${error}`);
-			return {animations: undefined};
+			return { animations: undefined };
 		}
 	}
 }
