@@ -1,9 +1,9 @@
-import {stdout} from 'process';
-import {Sustainability} from '.';
+import { Sustainability } from '.';
 
 (async () => {
 	Sustainability.auditStream.pipe(process.stdout);
 	await Sustainability.audit('https://audits.digital', {
-		connectionSettings: {streams: true}
+		connectionSettings: { streams: false, coldRun: false }
 	});
+
 })();
